move HttpClientProviderInterface in Provider namespace
